Refactor requirements modules
