+++ src/game/requirements.js
@@ -1,880 +1,64 @@
-import { DEFAULT_DAY_HOURS } from '../core/constants.js';
-import { formatDays, formatHours, formatList, formatMoney, toNumber } from '../core/helpers.js';
+import { getState } from '../core/state.js';
 import { addLog } from '../core/log.js';
-import {
-  countActiveAssetInstances,
-  getAssetState,
-  getState,
-  getUpgradeState
-} from '../core/state.js';
-import {
-  getAssetDefinition,
-  getUpgradeDefinition
-} from '../core/state/registry.js';
 import { spendMoney } from './currency.js';
 import { spendTime } from './time.js';
-import {
-  recordCostContribution,
-  recordTimeContribution
-} from './metrics.js';
-import { buildRequirementBundle, resolveRequirementConfig } from './schema/requirements.js';
+import { recordCostContribution, recordTimeContribution } from './metrics.js';
 import { awardSkillProgress } from './skills/index.js';
-import { ASSISTANT_CONFIG, getAssistantCount } from './assistant.js';
+import knowledgeTracks, { KNOWLEDGE_REWARDS, KNOWLEDGE_TRACKS } from './requirements/knowledgeTracks.js';
+import { getKnowledgeProgress } from './requirements/knowledgeProgress.js';
+import { estimateManualMaintenanceReserve } from './requirements/maintenanceReserve.js';
+import { getDefinitionRequirements } from './requirements/definitionRequirements.js';
+import {
+  buildAssetRequirementDescriptor,
+  describeRequirement,
+  formatAssetRequirementLabel,
+  summarizeAssetRequirements,
+  renderAssetRequirementDetail,
+  listAssetRequirementDescriptors,
+  updateAssetCardLock
+} from './requirements/descriptors.js';
+import {
+  isRequirementMet,
+  definitionRequirementsMet,
+  assetRequirementsMetById
+} from './requirements/checks.js';
+import { createRequirementsOrchestrator } from './requirements/orchestrator.js';
 
-const MIN_MANUAL_BUFFER_HOURS = Math.max(2, Math.round(DEFAULT_DAY_HOURS * 0.25));
+const orchestrator = createRequirementsOrchestrator({
+  getState,
+  getKnowledgeProgress,
+  knowledgeTracks,
+  knowledgeRewards: KNOWLEDGE_REWARDS,
+  estimateMaintenanceReserve: estimateManualMaintenanceReserve,
+  spendMoney,
+  spendTime,
+  recordCostContribution,
+  recordTimeContribution,
+  awardSkillProgress,
+  addLog
+});
 
-function estimateManualMaintenanceReserve(state) {
-  if (!state) return 0;
-  const assets = state.assets || {};
-  let upkeepDemand = 0;
-  let assistantHoursRemaining = getAssistantCount(state) * ASSISTANT_CONFIG.hoursPerAssistant;
+export const {
+  enrollInKnowledgeTrack,
+  dropKnowledgeTrack,
+  allocateDailyStudy,
+  advanceKnowledgeTracks
+} = orchestrator;
 
-  for (const [assetId, assetState] of Object.entries(assets)) {
-    if (!assetState?.instances?.length) continue;
-    const definition = getAssetDefinition(assetId);
-    if (!definition) continue;
-    const maintenanceHours = Number(definition.maintenance?.hours) || 0;
-    if (maintenanceHours <= 0) continue;
-
-    for (const instance of assetState.instances) {
-      if (instance?.status !== 'active' || instance?.maintenanceFundedToday) continue;
-
-      const assistantHoursApplied = Math.min(assistantHoursRemaining, maintenanceHours);
-      assistantHoursRemaining -= assistantHoursApplied;
-
-      const manualHoursNeeded = Math.max(0, maintenanceHours - assistantHoursApplied);
-      upkeepDemand += manualHoursNeeded;
-    }
-  }
-
-  return Math.max(0, upkeepDemand);
-}
-
-export function buildAssetRequirementDescriptor(requirement, state = getState(), typeOverride) {
-  if (!requirement || !requirement.assetId) {
-    return {
-      type: typeOverride ?? requirement?.type ?? 'unknown',
-      label: 'Unknown requirement',
-      met: false
-    };
-  }
-
-  const assetDefinition = getAssetDefinition(requirement.assetId);
-  const baseLabel = assetDefinition?.singular || assetDefinition?.name || requirement.assetId;
-  const parsedCount = Number(requirement.count);
-  const defaultNeed = requirement?.type === 'experience' ? 0 : 1;
-  const normalizedNeed = Number.isFinite(parsedCount)
-    ? Math.max(0, parsedCount)
-    : defaultNeed;
-  const need = normalizedNeed || defaultNeed;
-  const have = countActiveAssetInstances(requirement.assetId, state);
-
-  return {
-    type: typeOverride ?? requirement?.type ?? 'experience',
-    assetId: requirement.assetId,
-    label: `${need} ${baseLabel}${need === 1 ? '' : 's'}`,
-    met: have >= need,
-    progress: { have, need }
-  };
-}
-
-export const KNOWLEDGE_TRACKS = {
-  outlineMastery: {
-    id: 'outlineMastery',
-    name: 'Outline Mastery Workshop',
-    description: 'Deep-dive into narrative scaffolding for 5 days (2h/day). Tuition due upfront.',
-    hoursPerDay: 2,
-    days: 5,
-    tuition: 140,
-    instantBoosts: [
-      {
-        hustleId: 'freelance',
-        hustleName: 'Freelance Writing',
-        type: 'multiplier',
-        amount: 0.25
-      },
-      {
-        hustleId: 'audiobookNarration',
-        hustleName: 'Audiobook Narration Gig',
-        type: 'multiplier',
-        amount: 0.15
-      }
-    ]
-  },
-  photoLibrary: {
-    id: 'photoLibrary',
-    name: 'Photo Catalog Curation',
-    description: 'Archive, tag, and light-edit your best work for 4 days (1.5h/day). Tuition due upfront.',
-    hoursPerDay: 1.5,
-    days: 4,
-    tuition: 95,
-    instantBoosts: [
-      {
-        hustleId: 'eventPhotoGig',
-        hustleName: 'Event Photo Gig',
-        type: 'multiplier',
-        amount: 0.2
-      }
-    ]
-  },
-  ecomPlaybook: {
-    id: 'ecomPlaybook',
-    name: 'E-Commerce Playbook',
-    description: 'Shadow a pro operator for 7 days (2h/day) to master funnels and fulfillment math.',
-    hoursPerDay: 2,
-    days: 7,
-    tuition: 260,
-    instantBoosts: [
-      {
-        hustleId: 'bundlePush',
-        hustleName: 'Bundle Promo Push',
-        type: 'flat',
-        amount: 6
-      },
-      {
-        hustleId: 'dropshipPackParty',
-        hustleName: 'Dropship Pack Party',
-        type: 'multiplier',
-        amount: 0.2
-      }
-    ]
-  },
-  automationCourse: {
-    id: 'automationCourse',
-    name: 'Automation Architecture Course',
-    description: 'Pair-program with mentors for 10 days (~2¼h/day) to architect a reliable micro-app.',
-    hoursPerDay: 2.25,
-    days: 10,
-    tuition: 540,
-    instantBoosts: [
-      {
-        hustleId: 'saasBugSquash',
-        hustleName: 'SaaS Bug Squash',
-        type: 'flat',
-        amount: 12
-      }
-    ]
-  },
-  brandVoiceLab: {
-    id: 'brandVoiceLab',
-    name: 'Brand Voice Lab',
-    description: 'Work with pitch coaches for 4 days (1h/day) to sharpen live Q&A charisma.',
-    hoursPerDay: 1,
-    days: 4,
-    tuition: 120,
-    instantBoosts: [
-      {
-        hustleId: 'audienceCall',
-        hustleName: 'Audience Q&A Blast',
-        type: 'flat',
-        amount: 4
-      }
-    ]
-  },
-  guerillaBuzzWorkshop: {
-    id: 'guerillaBuzzWorkshop',
-    name: 'Guerrilla Buzz Workshop',
-    description: 'Field-test hype hooks for 6 days (1.5h/day) with a crew of street marketers.',
-    hoursPerDay: 1.5,
-    days: 6,
-    tuition: 180,
-    instantBoosts: [
-      {
-        hustleId: 'streetPromoSprint',
-        hustleName: 'Street Promo Sprint',
-        type: 'multiplier',
-        amount: 0.25
-      },
-      {
-        hustleId: 'surveySprint',
-        hustleName: 'Micro Survey Dash',
-        type: 'flat',
-        amount: 1.5
-      }
-    ]
-  },
-  curriculumDesignStudio: {
-    id: 'curriculumDesignStudio',
-    name: 'Curriculum Design Studio',
-    description: 'Prototype interactive lesson plans for 6 days (2.5h/day) with veteran educators.',
-    hoursPerDay: 2.5,
-    days: 6,
-    tuition: 280,
-    instantBoosts: [
-      {
-        hustleId: 'popUpWorkshop',
-        hustleName: 'Pop-Up Workshop',
-        type: 'multiplier',
-        amount: 0.3
-      },
-      {
-        hustleId: 'bundlePush',
-        hustleName: 'Bundle Promo Push',
-        type: 'multiplier',
-        amount: 0.15
-      }
-    ]
-  },
-  postProductionPipelineLab: {
-    id: 'postProductionPipelineLab',
-    name: 'Post-Production Pipeline Lab',
-    description: 'Run edit bays for 8 days (3h/day) to master color, captions, and delivery workflows.',
-    hoursPerDay: 3,
-    days: 8,
-    tuition: 360,
-    instantBoosts: [
-      {
-        hustleId: 'vlogEditRush',
-        hustleName: 'Vlog Edit Rush',
-        type: 'multiplier',
-        amount: 0.35
-      },
-      {
-        assetId: 'vlog',
-        assetName: 'Weekly Vlog Channel',
-        type: 'multiplier',
-        amount: 0.18
-      }
-    ]
-  },
-  fulfillmentOpsMasterclass: {
-    id: 'fulfillmentOpsMasterclass',
-    name: 'Fulfillment Ops Masterclass',
-    description: 'Shadow a 7-day (2h/day) logistics crew to automate pick, pack, and ship perfection.',
-    hoursPerDay: 2,
-    days: 7,
-    tuition: 320,
-    instantBoosts: [
-      {
-        hustleId: 'dropshipPackParty',
-        hustleName: 'Dropship Pack Party',
-        type: 'multiplier',
-        amount: 0.25
-      },
-      {
-        assetId: 'dropshipping',
-        assetName: 'Dropshipping Product Lab',
-        type: 'multiplier',
-        amount: 0.35
-      }
-    ]
-  },
-  customerRetentionClinic: {
-    id: 'customerRetentionClinic',
-    name: 'Customer Retention Clinic',
-    description: 'Coach subscription success teams for 5 days (2h/day) to keep churn near zero.',
-    hoursPerDay: 2,
-    days: 5,
-    tuition: 210,
-    instantBoosts: [
-      {
-        hustleId: 'saasBugSquash',
-        hustleName: 'SaaS Bug Squash',
-        type: 'flat',
-        amount: 8
-      },
-      {
-        assetId: 'saas',
-        assetName: 'SaaS Micro-App',
-        type: 'multiplier',
-        amount: 0.25
-      }
-    ]
-  },
-  narrationPerformanceWorkshop: {
-    id: 'narrationPerformanceWorkshop',
-    name: 'Narration Performance Workshop',
-    description: 'Spend 4 days (1.75h/day) with vocal coaches to perfect audiobook cadence.',
-    hoursPerDay: 1.75,
-    days: 4,
-    tuition: 190,
-    instantBoosts: [
-      {
-        hustleId: 'audiobookNarration',
-        hustleName: 'Audiobook Narration',
-        type: 'multiplier',
-        amount: 0.3
-      },
-      {
-        assetId: 'ebook',
-        assetName: 'Digital E-Book Series',
-        type: 'multiplier',
-        amount: 0.15
-      }
-    ]
-  },
-  galleryLicensingSummit: {
-    id: 'galleryLicensingSummit',
-    name: 'Gallery Licensing Summit',
-    description: 'Pitch curators for 5 days (2.25h/day) to secure premium gallery licensing deals.',
-    hoursPerDay: 2.25,
-    days: 5,
-    tuition: 240,
-    instantBoosts: [
-      {
-        hustleId: 'eventPhotoGig',
-        hustleName: 'Event Photo Gig',
-        type: 'multiplier',
-        amount: 0.3
-      },
-      {
-        assetId: 'stockPhotos',
-        assetName: 'Stock Photo Gallery',
-        type: 'multiplier',
-        amount: 0.22
-      }
-    ]
-  },
-  syndicationResidency: {
-    id: 'syndicationResidency',
-    name: 'Syndication Residency',
-    description: 'Curate partnerships for 6 days (2h/day) to syndicate your flagship content network.',
-    hoursPerDay: 2,
-    days: 6,
-    tuition: 300,
-    instantBoosts: [
-      {
-        hustleId: 'freelance',
-        hustleName: 'Freelance Writing',
-        type: 'multiplier',
-        amount: 0.2
-      },
-      {
-        hustleId: 'streetPromoSprint',
-        hustleName: 'Street Promo Sprint',
-        type: 'flat',
-        amount: 2
-      },
-      {
-        assetId: 'blog',
-        assetName: 'Personal Blog Network',
-        type: 'multiplier',
-        amount: 0.18
-      }
-    ]
-  }
+export {
+  KNOWLEDGE_TRACKS,
+  KNOWLEDGE_REWARDS,
+  getKnowledgeProgress,
+  estimateManualMaintenanceReserve,
+  getDefinitionRequirements,
+  buildAssetRequirementDescriptor,
+  describeRequirement,
+  formatAssetRequirementLabel,
+  summarizeAssetRequirements,
+  renderAssetRequirementDetail,
+  listAssetRequirementDescriptors,
+  updateAssetCardLock,
+  isRequirementMet,
+  definitionRequirementsMet,
+  assetRequirementsMetById
 };
-
-export const KNOWLEDGE_REWARDS = {
-  outlineMastery: { baseXp: 120, skills: ['writing'] },
-  photoLibrary: {
-    baseXp: 120,
-    skills: [
-      { id: 'visual', weight: 0.5 },
-      { id: 'editing', weight: 0.5 }
-    ]
-  },
-  ecomPlaybook: {
-    baseXp: 120,
-    skills: [
-      { id: 'research', weight: 0.5 },
-      { id: 'commerce', weight: 0.5 }
-    ]
-  },
-  automationCourse: {
-    baseXp: 120,
-    skills: [
-      { id: 'software', weight: 0.6 },
-      { id: 'infrastructure', weight: 0.4 }
-    ]
-  },
-  brandVoiceLab: {
-    baseXp: 100,
-    skills: [
-      { id: 'audience', weight: 0.6 },
-      { id: 'promotion', weight: 0.4 }
-    ]
-  },
-  guerillaBuzzWorkshop: {
-    baseXp: 110,
-    skills: [
-      { id: 'promotion', weight: 0.6 },
-      { id: 'audience', weight: 0.4 }
-    ]
-  },
-  curriculumDesignStudio: {
-    baseXp: 150,
-    skills: [
-      { id: 'audience', weight: 0.6 },
-      { id: 'writing', weight: 0.4 }
-    ]
-  },
-  postProductionPipelineLab: {
-    baseXp: 150,
-    skills: [
-      { id: 'editing', weight: 0.7 },
-      { id: 'visual', weight: 0.3 }
-    ]
-  },
-  fulfillmentOpsMasterclass: {
-    baseXp: 140,
-    skills: [
-      { id: 'commerce', weight: 0.7 },
-      { id: 'promotion', weight: 0.3 }
-    ]
-  },
-  customerRetentionClinic: {
-    baseXp: 140,
-    skills: [
-      { id: 'audience', weight: 0.4 },
-      { id: 'promotion', weight: 0.3 },
-      { id: 'software', weight: 0.3 }
-    ]
-  },
-  narrationPerformanceWorkshop: {
-    baseXp: 130,
-    skills: [
-      { id: 'audio', weight: 0.6 },
-      { id: 'writing', weight: 0.4 }
-    ]
-  },
-  galleryLicensingSummit: {
-    baseXp: 140,
-    skills: [
-      { id: 'visual', weight: 0.6 },
-      { id: 'commerce', weight: 0.4 }
-    ]
-  },
-  syndicationResidency: {
-    baseXp: 150,
-    skills: [
-      { id: 'promotion', weight: 0.5 },
-      { id: 'audience', weight: 0.3 },
-      { id: 'writing', weight: 0.2 }
-    ]
-  }
-};
-
-const EMPTY_REQUIREMENTS = buildRequirementBundle();
-const requirementCache = new WeakMap();
-
-export function getDefinitionRequirements(definition) {
-  if (!definition) return EMPTY_REQUIREMENTS;
-  if (requirementCache.has(definition)) {
-    return requirementCache.get(definition);
-  }
-  const config = resolveRequirementConfig(definition);
-  if (!config) {
-    requirementCache.set(definition, EMPTY_REQUIREMENTS);
-    return EMPTY_REQUIREMENTS;
-  }
-  const bundle = buildRequirementBundle(config);
-  requirementCache.set(definition, bundle);
-  return bundle;
-}
-
-function isEquipmentUnlocked(id, state = getState()) {
-  if (!id) return true;
-  return Boolean(getUpgradeState(id, state).purchased);
-}
-
-function isKnowledgeComplete(id, state = getState()) {
-  if (!id) return true;
-  const track = KNOWLEDGE_TRACKS[id];
-  if (!track) return true;
-  const progress = getKnowledgeProgress(id, state);
-  return progress.completed;
-}
-
-function hasExperience(requirement, state = getState()) {
-  if (!requirement?.assetId) return true;
-  const targetCount = Number(requirement.count) || 0;
-  if (targetCount <= 0) return true;
-  const owned = countActiveAssetInstances(requirement.assetId, state);
-  return owned >= targetCount;
-}
-
-export function isRequirementMet(requirement, state = getState()) {
-  if (!requirement) return true;
-  switch (requirement.type) {
-    case 'equipment':
-      return isEquipmentUnlocked(requirement.id, state);
-    case 'knowledge':
-      return isKnowledgeComplete(requirement.id, state);
-    case 'experience':
-      return hasExperience(requirement, state);
-    default:
-      return true;
-  }
-}
-
-export function describeRequirement(requirement, state = getState()) {
-  if (!requirement) {
-    return {
-      type: 'unknown',
-      status: 'unknown',
-      icon: '❓',
-      label: 'Unknown Requirement',
-      detail: '❓ <strong>Unknown requirement</strong>'
-    };
-  }
-
-  const status = isRequirementMet(requirement, state) ? 'met' : 'pending';
-
-  if (requirement.type === 'equipment') {
-    const upgrade = getUpgradeDefinition(requirement.id);
-    const label = upgrade?.name || requirement.id;
-    const icon = status === 'met' ? '✅' : '🔒';
-    return {
-      type: 'equipment',
-      status,
-      icon,
-      label,
-      detail: `${icon} <strong>${label}</strong>`
-    };
-  }
-
-  if (requirement.type === 'knowledge') {
-    const track = KNOWLEDGE_TRACKS[requirement.id];
-    const label = track?.name || requirement.id;
-    const progress = getKnowledgeProgress(requirement.id, state);
-    const icon = progress.completed ? '✅' : progress.studiedToday ? '📗' : '📘';
-    const hoursPerDay = Number(track?.hoursPerDay) || 0;
-    const detail = track
-      ? `${progress.daysCompleted}/${track.days} days, ${formatHours(hoursPerDay)}/day`
-      : 'Progress tracked';
-    return {
-      type: 'knowledge',
-      status,
-      icon,
-      label,
-      detail: `${icon} <strong>${label}</strong> (${detail})`
-    };
-  }
-
-  if (requirement.type === 'experience') {
-    const assetDef = getAssetDefinition(requirement.assetId);
-    const owned = countActiveAssetInstances(requirement.assetId, state);
-    const target = Number(requirement.count) || 0;
-    const baseLabel = assetDef?.singular || assetDef?.name || requirement.assetId;
-    const label = `${target} ${baseLabel}${target === 1 ? '' : 's'}`;
-    const icon = owned >= target ? '✅' : '🏆';
-    return {
-      type: 'experience',
-      status,
-      icon,
-      label,
-      detail: `${icon} <strong>${label}</strong> (have ${owned})`
-    };
-  }
-
-  return {
-    type: requirement.type,
-    status,
-    icon: status === 'met' ? '✅' : '❔',
-    label: 'Unknown Requirement',
-    detail: `${status === 'met' ? '✅' : '❔'} <strong>Unknown requirement</strong>`
-  };
-}
-
-export function definitionRequirementsMet(definition, state = getState()) {
-  const requirements = getDefinitionRequirements(definition);
-  if (!requirements.hasAny) return true;
-  return requirements.every(req => isRequirementMet(req, state));
-}
-
-export function assetRequirementsMetById(id, state = getState()) {
-  const definition = getAssetDefinition(id);
-  if (!definition) return true;
-  return definitionRequirementsMet(definition, state);
-}
-
-export function formatAssetRequirementLabel(assetId, state = getState()) {
-  const definition = getAssetDefinition(assetId);
-  if (!definition) return 'Requirement Missing';
-  const requirements = getDefinitionRequirements(definition);
-  if (!requirements.hasAny) return 'Ready to Launch';
-  const missing = requirements.missing(req => isRequirementMet(req, state));
-  if (!missing.length) return 'Ready to Launch';
-  const names = missing.map(req => describeRequirement(req, state).label);
-  return `Requires ${names.join(' & ')}`;
-}
-
-export function summarizeAssetRequirements(requirements = [], state = getState()) {
-  if (!requirements?.length) return 'None';
-  return requirements
-    .map(req => {
-      const definition = getAssetDefinition(req.assetId);
-      const label = definition?.singular || definition?.name || req.assetId;
-      const need = toNumber(req.count, 1);
-      const have = countActiveAssetInstances(req.assetId, state);
-      return `${label}: ${have}/${need} active`;
-    })
-    .join(' • ');
-}
-
-export function renderAssetRequirementDetail(assetId, state = getState()) {
-  const definition = getAssetDefinition(assetId);
-  if (!definition) return '';
-  const requirements = getDefinitionRequirements(definition);
-  if (!requirements.hasAny) {
-    return '🔓 Requirements: <strong>None</strong>';
-  }
-  const parts = requirements.map(req => describeRequirement(req, state).detail);
-  return `Requirements: ${parts.join(' • ')}`;
-}
-
-export function listAssetRequirementDescriptors(definitionOrId, state = getState()) {
-  const definition = typeof definitionOrId === 'string'
-    ? getAssetDefinition(definitionOrId)
-    : definitionOrId;
-  if (!definition) return [];
-  const requirements = getDefinitionRequirements(definition);
-  if (!requirements.hasAny) return [];
-
-  return requirements.map(req => {
-    switch (req.type) {
-      case 'equipment': {
-        const upgrade = getUpgradeDefinition(req.id);
-        return {
-          type: 'equipment',
-          id: req.id,
-          label: upgrade?.name || req.id,
-          met: isEquipmentUnlocked(req.id, state)
-        };
-      }
-      case 'knowledge': {
-        const track = KNOWLEDGE_TRACKS[req.id];
-        const progress = getKnowledgeProgress(req.id, state);
-        return {
-          type: 'knowledge',
-          id: req.id,
-          label: track?.name || req.id,
-          met: progress.completed,
-          progress: {
-            daysCompleted: progress.daysCompleted,
-            totalDays: track?.days ?? 0,
-            studiedToday: Boolean(progress.studiedToday)
-          }
-        };
-      }
-      case 'experience': {
-        return buildAssetRequirementDescriptor(req, state);
-      }
-      default:
-        return { type: req.type || 'unknown', label: 'Unknown requirement', met: false };
-    }
-  });
-}
-
-export function updateAssetCardLock(assetId, card, state = getState()) {
-  const definition = getAssetDefinition(assetId);
-  if (!definition || !card) return;
-  const requirements = getDefinitionRequirements(definition);
-  const locked = requirements.hasAny && !requirements.every(req => isRequirementMet(req, state));
-  card.classList.toggle('locked', locked);
-}
-
-export function getKnowledgeProgress(id, target = getState()) {
-  target.progress = target.progress || {};
-  target.progress.knowledge = target.progress.knowledge || {};
-  if (!target.progress.knowledge[id]) {
-    const track = KNOWLEDGE_TRACKS[id];
-    target.progress.knowledge[id] = {
-      daysCompleted: 0,
-      studiedToday: false,
-      completed: false,
-      enrolled: false,
-      totalDays: track?.days ?? 0,
-      hoursPerDay: track?.hoursPerDay ?? 0,
-      tuitionCost: track?.tuition ?? 0,
-      enrolledOnDay: null,
-      skillRewarded: false
-    };
-  }
-  const track = KNOWLEDGE_TRACKS[id];
-  const progress = target.progress.knowledge[id];
-  if (track) {
-    progress.totalDays = track.days;
-    progress.hoursPerDay = track.hoursPerDay;
-    progress.tuitionCost = track.tuition ?? 0;
-    progress.completed = progress.completed || progress.daysCompleted >= track.days;
-  }
-  progress.skillRewarded = Boolean(progress.skillRewarded);
-  return progress;
-}
-
-export function enrollInKnowledgeTrack(id) {
-  const state = getState();
-  const track = KNOWLEDGE_TRACKS[id];
-  if (!state || !track) return { success: false, reason: 'missing' };
-
-  const progress = getKnowledgeProgress(id);
-  if (progress.completed) {
-    addLog(`${track.name} is already complete. Grab a celebratory pastry instead.`, 'info');
-    return { success: false, reason: 'completed' };
-  }
-  if (progress.enrolled) {
-    addLog(`You're already enrolled in ${track.name}.`, 'info');
-    return { success: false, reason: 'enrolled' };
-  }
-
-  const tuition = Number(track.tuition) || 0;
-  if (tuition > 0 && state.money < tuition) {
-    addLog(`You need $${formatMoney(tuition)} ready to enroll in ${track.name}.`, 'warning');
-    return { success: false, reason: 'money' };
-  }
-
-  if (tuition > 0) {
-    spendMoney(tuition);
-    recordCostContribution({
-      key: `study:${track.id}:tuition`,
-      label: `🎓 ${track.name} tuition`,
-      amount: tuition,
-      category: 'investment'
-    });
-  }
-
-  progress.enrolled = true;
-  progress.enrolledOnDay = state.day;
-  progress.studiedToday = false;
-
-  addLog(`You enrolled in ${track.name}! Tuition paid${tuition > 0 ? ` for $${formatMoney(tuition)}` : ''}.`, 'info');
-
-  allocateDailyStudy({ trackIds: [id], triggeredByEnrollment: true });
-
-  return { success: true };
-}
-
-export function dropKnowledgeTrack(id) {
-  const state = getState();
-  const track = KNOWLEDGE_TRACKS[id];
-  if (!state || !track) {
-    return { success: false, reason: 'missing' };
-  }
-
-  const progress = getKnowledgeProgress(id, state);
-  if (!progress.enrolled || progress.completed) {
-    addLog(`You're not currently enrolled in ${track.name}.`, 'info');
-    return { success: false, reason: 'not_enrolled' };
-  }
-
-  progress.enrolled = false;
-  progress.studiedToday = false;
-  progress.enrolledOnDay = null;
-
-  addLog(`You dropped ${track.name}. Tuition stays paid, but your schedule opens back up.`, 'warning');
-
-  return { success: true };
-}
-
-export function allocateDailyStudy({ trackIds, triggeredByEnrollment = false } = {}) {
-  const state = getState();
-  if (!state) return;
-
-  const studied = [];
-  const reserveSkipped = [];
-  const timeSkipped = [];
-
-  const tracks = trackIds
-    ? trackIds.map(id => KNOWLEDGE_TRACKS[id]).filter(Boolean)
-    : Object.values(KNOWLEDGE_TRACKS);
-
-  const maintenanceReserve = estimateManualMaintenanceReserve(state);
-  let availableStudyTime = Math.max(0, state.timeLeft - maintenanceReserve - MIN_MANUAL_BUFFER_HOURS);
-
-  for (const track of tracks) {
-    const progress = getKnowledgeProgress(track.id);
-    if (!progress.enrolled || progress.completed) continue;
-    if (progress.studiedToday) continue;
-
-    const hours = Number(track.hoursPerDay) || 0;
-    if (hours <= 0) {
-      progress.studiedToday = true;
-      continue;
-    }
-
-    if (availableStudyTime < hours) {
-      reserveSkipped.push(track.name);
-      continue;
-    }
-
-    if (state.timeLeft < hours) {
-      timeSkipped.push(track.name);
-      continue;
-    }
-
-    spendTime(hours);
-    availableStudyTime = Math.max(0, availableStudyTime - hours);
-    recordTimeContribution({
-      key: `study:${track.id}:time`,
-      label: `📘 ${track.name} study`,
-      hours,
-      category: 'study'
-    });
-    progress.studiedToday = true;
-    studied.push(track.name);
-  }
-
-  if (studied.length) {
-    const prefix = triggeredByEnrollment ? 'Class time booked today for' : 'Study sessions reserved for';
-    addLog(`${prefix} ${formatList(studied)}.`, 'info');
-  }
-
-  if (reserveSkipped.length) {
-    addLog(
-      `${formatList(reserveSkipped)} were deferred to leave breathing room for upkeep commitments.`,
-      'warning'
-    );
-  }
-
-  if (timeSkipped.length) {
-    addLog(`${formatList(timeSkipped)} could not fit into today\'s schedule.`, 'warning');
-  }
-}
-
-export function advanceKnowledgeTracks() {
-  const state = getState();
-  if (!state) return;
-
-  const completedToday = [];
-  const stalled = [];
-
-  Object.entries(state.progress.knowledge || {}).forEach(([id, progress]) => {
-    const track = KNOWLEDGE_TRACKS[id];
-    if (!track) {
-      progress.studiedToday = false;
-      return;
-    }
-    if (progress.completed) {
-      progress.studiedToday = false;
-      return;
-    }
-
-    if (!progress.enrolled) {
-      progress.studiedToday = false;
-      return;
-    }
-
-    if (progress.studiedToday) {
-      const before = progress.daysCompleted;
-      progress.daysCompleted = Math.min(track.days, before + 1);
-      progress.studiedToday = false;
-      if (progress.daysCompleted >= track.days) {
-        progress.completed = true;
-        progress.enrolled = false;
-        if (!progress.skillRewarded) {
-          const reward = KNOWLEDGE_REWARDS[id];
-          if (reward) {
-            awardSkillProgress({
-              skills: reward.skills,
-              baseXp: reward.baseXp,
-              label: track.name
-            });
-          }
-          progress.skillRewarded = true;
-        }
-        completedToday.push(track.name);
-      }
-    } else if (progress.daysCompleted > 0 || progress.enrolled) {
-      stalled.push(track.name);
-    }
-  });
-
-  if (completedToday.length) {
-    addLog(`Finished ${formatList(completedToday)}! New opportunities unlocked.`, 'info');
-  }
-  if (stalled.length) {
-    addLog(`${formatList(stalled)} did not get study time today. Progress paused.`, 'warning');
-  }
-}
